requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
dash==2.0.0
dash-auth==1.3.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.0
dash-table==5.0.0
dash-uploader==0.6.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-SeaSurf==0.3.1
future==0.18.2
gunicorn==20.1.0
h5py==3.5.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
llvmlite==0.37.0
loompy==3.0.6
MarkupSafe==2.0.1
numba==0.54.1
numpy==1.20.3
numpy-groupies==0.9.14
packaging==21.2
pandas==1.3.4
plotly==5.1.0
py-healthcheck==1.10.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.3
requests==2.26.0
retrying==1.3.3
scipy==1.7.1
setuptools==58.0.4
six==1.15.0
tenacity==8.0.1
ua-parser==0.10.0
urllib3==1.26.7
Werkzeug==2.0.1
wheel==0.37.0
wincertstore==0.2
zipp==3.6.0
